requirements.txt
fastapi
uvicorn
pydantic
requests
httpx
pandas
prophet
psycopg2-binary
langchain-core
langchain-ollama

